backend/requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.20
django-chamber==0.6.17
django-choice-enumfields==1.1.3
django-generic-m2m-field==0.0.9
django-ipware==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-yasg==1.21.7
eventlet==0.33.3
filemagic==1.6
flake8==5.0.4
greenlet==2.0.2
idna==3.4
import-string==0.1.0
importlib-metadata==4.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.1
prompt-toolkit==3.0.39
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.8.0
PyPrind==2.11.3
pytz==2023.3
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.0
typing_extensions==4.7.1
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
